+++ claims_test.go
@@ -37,10 +37,10 @@ func TestGetClaim(t *testing.T) {
 		claim string
 		value interface{}
 	}{
-		{"a", float64(42)},
-		{"/a", float64(42)},
+		{"a", json.Number("42")},
+		{"/a", json.Number("42")},
 		{"b", "bar"},
-		{"/c/d", float64(95)},
+		{"/c/d", json.Number("95")},
 		{"/c/e/1", "cat"},
 		{"/c/f/g", "zebra"},
 		{"nope", nil},
@@ -83,10 +83,10 @@ func TestSetClaim(t *testing.T) {
 		claim string
 		value interface{}
 	}{
-		{"a", float64(43)},
-		{"/a", float64(43)},
+		{"a", json.Number("43")},
+		{"/a", json.Number("43")},
 		{"b", "foo"},
-		{"/c/d", float64(96)},
+		{"/c/d", json.Number("96")},
 		{"/c/e/1", "dog"},
 		{"/c/f/g", "elephant"},
 	}
@@ -290,6 +290,54 @@ func TestValidateBoundClaims(t *testing.T) {
 				"foo": []interface{}{"c", "d"},
 			},
 			errExpected: false,
+		},
+		{
+			name:            "valid match with numeric claim conversion from float64",
+			boundClaimsType: "string",
+			boundClaims: map[string]interface{}{
+				// Numeric bound claims from Vault API are json.Number type
+				"foo": json.Number("123"),
+			},
+			allClaims: map[string]interface{}{
+				"foo": float64(123),
+			},
+			errExpected: false,
+		},
+		{
+			name:            "valid match with numeric claim conversion from float32",
+			boundClaimsType: "string",
+			boundClaims: map[string]interface{}{
+				// Numeric bound claims from Vault API are json.Number type
+				"foo": json.Number("123"),
+			},
+			allClaims: map[string]interface{}{
+				"foo": float32(123),
+			},
+			errExpected: false,
+		},
+		{
+			name:            "invalid match with numeric claim conversion from float64",
+			boundClaimsType: "string",
+			boundClaims: map[string]interface{}{
+				// Numeric bound claims from Vault API are json.Number type
+				"foo": json.Number("456"),
+			},
+			allClaims: map[string]interface{}{
+				"foo": float64(123),
+			},
+			errExpected: true,
+		},
+		{
+			name:            "invalid match with numeric claim conversion from float32",
+			boundClaimsType: "string",
+			boundClaims: map[string]interface{}{
+				// Numeric bound claims from Vault API are json.Number type
+				"foo": json.Number("123"),
+			},
+			allClaims: map[string]interface{}{
+				"foo": float32(456),
+			},
+			errExpected: true,
 		},
 		{
 			name:            "invalid - no match within list",